+++ CHANGELOG.md
@@ -22,6 +22,9 @@ Fixes
 - Fixed StaticSpriteBatch handling rotated sprites incorrectly
 
 ### MLEM.Ui
+Additions
+- Added Element.OnStyleInit event
+
 Improvements
 - Allow for checkboxes and radio buttons to be disabled
 - Only set a paragraph's area dirty when a text change would cause it to change size